+++ src/util/entity.hpp
@@ -19,6 +19,6 @@ namespace big::entity
 	std::vector<Entity> get_entities(bool vehicles, bool peds, bool props = false, bool include_self_veh = false);
 	bool load_ground_at_3dcoord(Vector3& location);
 	bool request_model(rage::joaat_t hash);
-	double distance_to_middle_of_screen(const rage::fvector2& screen_pos);
+	float distance_to_middle_of_screen(const rage::fvector2& screen_pos);
 	Entity get_entity_closest_to_middle_of_screen(rage::fwEntity** pointer = nullptr, std::vector<Entity> ignore_entities = {}, bool include_veh = true, bool include_ped = true, bool include_prop = true, bool include_players = true);
 }
